requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
streamlit==1.29.0
scikit-learn==1.3.2